+++ src/MooVC.Syntax.Tests/Attributes/Resource/AssemblyTests/WhenInequalityOperatorAssemblyAssemblyIsCalled.cs
@@ -1,0 +1,62 @@
+namespace MooVC.Syntax.Attributes.Resource.AssemblyTests;
+
+using MooVC.Syntax.Elements;
+
+public sealed class WhenInequalityOperatorAssemblyAssemblyIsCalled
+{
+    [Fact]
+    public void GivenBothNullThenReturnsFalse()
+    {
+        // Arrange
+        Assembly? left = default;
+        Assembly? right = default;
+
+        // Act
+        bool result = left != right;
+
+        // Assert
+        result.ShouldBeFalse();
+    }
+
+    [Fact]
+    public void GivenLeftNullRightValueThenReturnsTrue()
+    {
+        // Arrange
+        Assembly? left = default;
+        Assembly right = AssemblyTestsData.Create();
+
+        // Act
+        bool result = left != right;
+
+        // Assert
+        result.ShouldBeTrue();
+    }
+
+    [Fact]
+    public void GivenEqualValuesThenReturnsFalse()
+    {
+        // Arrange
+        Assembly left = AssemblyTestsData.Create();
+        Assembly right = AssemblyTestsData.Create();
+
+        // Act
+        bool result = left != right;
+
+        // Assert
+        result.ShouldBeFalse();
+    }
+
+    [Fact]
+    public void GivenDifferentValuesThenReturnsTrue()
+    {
+        // Arrange
+        Assembly left = AssemblyTestsData.Create();
+        Assembly right = AssemblyTestsData.Create(name: Snippet.From("Other"));
+
+        // Act
+        bool result = left != right;
+
+        // Assert
+        result.ShouldBeTrue();
+    }
+}