+++ src/MooVC.Syntax.Tests/Attributes/Resource/HeaderTests/WhenEqualsObjectIsCalled.cs
@@ -1,0 +1,44 @@
+namespace MooVC.Syntax.Attributes.Resource.HeaderTests;
+
+public sealed class WhenEqualsObjectIsCalled
+{
+    [Fact]
+    public void GivenNullThenReturnsFalse()
+    {
+        // Arrange
+        Header subject = HeaderTestsData.Create();
+
+        // Act
+        bool result = subject.Equals(null);
+
+        // Assert
+        result.ShouldBeFalse();
+    }
+
+    [Fact]
+    public void GivenDifferentTypeThenReturnsFalse()
+    {
+        // Arrange
+        Header subject = HeaderTestsData.Create();
+
+        // Act
+        bool result = subject.Equals(new object());
+
+        // Assert
+        result.ShouldBeFalse();
+    }
+
+    [Fact]
+    public void GivenEqualValuesThenReturnsTrue()
+    {
+        // Arrange
+        Header subject = HeaderTestsData.Create();
+        object other = HeaderTestsData.Create();
+
+        // Act
+        bool result = subject.Equals(other);
+
+        // Assert
+        result.ShouldBeTrue();
+    }
+}