+++ src/MooVC.Syntax.Tests/Concepts/ResourceTests/WhenEqualsResourceIsCalled.cs
@@ -1,0 +1,48 @@
+namespace MooVC.Syntax.Concepts.ResourceTests;
+
+using MooVC.Syntax.Attributes.Resource;
+
+public sealed class WhenEqualsResourceIsCalled
+{
+    [Fact]
+    public void GivenNullThenReturnsFalse()
+    {
+        // Arrange
+        Resource subject = ResourceTestsData.Create();
+        Resource? other = default;
+
+        // Act
+        bool result = subject.Equals(other);
+
+        // Assert
+        result.ShouldBeFalse();
+    }
+
+    [Fact]
+    public void GivenEqualValuesThenReturnsTrue()
+    {
+        // Arrange
+        Resource subject = ResourceTestsData.Create();
+        Resource other = ResourceTestsData.Create();
+
+        // Act
+        bool result = subject.Equals(other);
+
+        // Assert
+        result.ShouldBeTrue();
+    }
+
+    [Fact]
+    public void GivenDifferentValuesThenReturnsFalse()
+    {
+        // Arrange
+        Resource subject = ResourceTestsData.Create();
+        Resource other = ResourceTestsData.Create(data: Data.Undefined);
+
+        // Act
+        bool result = subject.Equals(other);
+
+        // Assert
+        result.ShouldBeFalse();
+    }
+}