+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 
 - A CSharp Syntax generator based on the [Fluent Builder](https://github.com/MooVC/Fluentify) pattern.
+- Added a Resource concept for representing RESX files.
+- Added ProjectExtensions.Add for wiring RESX resources into project items.
 
 # [9.2.0] - 2025-11-14